refactor(fe): 상세페이지 리팩토링 (#135) (#137)
* Refactor(fe): 상세페이지 리팩토링

Components 폴더를 만들어서 상세페이지에 사용한 Component 들을 모두 보관

* Fix(be): 댓글 삭제 예외처리 추가

Comment->user가 null인 경우, 양방향 관계를 제거하지 않는다.
