+++ backend/src/main/java/com/back/ourlog/domain/comment/service/CommentService.java
@@ -58,7 +58,8 @@ public class CommentService {
         diary.deleteComment(comment);
 
         User user = comment.getUser();
-        user.deleteComment(comment);
+
+        if(user != null) user.deleteComment(comment);
 
         commentRepository.delete(comment);
     }
